+++ catwalk/cached_transformers.py
@@ -163,6 +163,9 @@ def get_tokenizer(cls: Type[T], model_name: str, **kwargs) -> T:
         # This issue will be fixed soon, see: https://github.com/huggingface/tokenizers/pull/1005. so that the fast tokenizer works correctly.  
         if model_name.startswith('facebook/opt'):
             kwargs['use_fast'] = False
+        elif model_name.startswith('t5-'):
+            # Workaround for another huggingface tokenizer bug.
+            kwargs['model_max_length'] = int(1e30)
         tokenizer = cls.from_pretrained(  # type: ignore
             model_name,
             **kwargs,