+++ catwalk/tasks/eleuther.py
@@ -3,7 +3,8 @@ from typing import Dict, Any, Optional, Union, Callable, Sequence, List, TypeVar
 
 from tango.common.sequences import MappedSequence
 
-from catwalk.task import Task, InstanceFormat, RankClassificationInstance
+from catwalk.task import Task, InstanceFormat, RankClassificationInstance, WithAnswerOptionsMixin, \
+    classification_metrics
 
 import lm_eval.tasks
 from lm_eval.base import Task as EAITask
@@ -140,6 +141,20 @@ class EleutherTask(Task):
         return RankClassificationInstance(choices, label)
 
 
+@Task.register("eleuther::classification")
+class EleutherClassificationTask(EleutherTask, WithAnswerOptionsMixin):
+    def __init__(
+        self,
+        eleuther_task: Union[str, Callable[[], EAITask]],
+        *,
+        answer_options: Sequence[str],
+        version_override: Optional[str] = None,
+    ):
+        EleutherTask.__init__(self, eleuther_task, version_override=version_override, ranked_classification=True)
+        WithAnswerOptionsMixin.__init__(self, answer_options)
+        self.add_metrics(classification_metrics(len(answer_options)))
+
+
 @Task.register("eleuther::race")
 class RaceEleutherTask(EleutherTask):
     """This task is different because there is no 1:1 correspondence between HF instances and EAI instances."""
@@ -168,11 +183,20 @@ class RaceEleutherTask(EleutherTask):
         raise KeyError(split)
 
 
-@Task.register("eleuther::pubmedqa")
-class PubmedqaEleutherTask(EleutherTask):
+@Task.register("eleuther::renamed_splits")
+class EleutherTaskWithRenamedSplits(EleutherTask):
     """This task is different because EAI relabels the datasets."""
-    def __init__(self, *, version_override: Optional[str] = None):
-        super().__init__("pubmedqa", version_override=version_override)
+    def __init__(
+        self,
+        eleuther_task: Union[str, Callable[[], EAITask]],
+        *,
+        version_override: Optional[str] = None,
+        ranked_classification: bool = False
+    ):
+        super().__init__(
+            eleuther_task,
+            version_override=version_override,
+            ranked_classification=ranked_classification)
 
     def has_split(self, split: str) -> bool:
         if split == "train":
@@ -196,3 +220,20 @@ class PubmedqaEleutherTask(EleutherTask):
         # to make them act like sequences.
         return MappedSequence(lambda x: x, result)
 
+
+@Task.register("eleuther::classification_with_renamed_splits")
+class EleutherClassificationTaskWithRenamedSplits(EleutherTaskWithRenamedSplits, WithAnswerOptionsMixin):
+    def __init__(
+        self,
+        eleuther_task: Union[str, Callable[[], EAITask]],
+        *,
+        answer_options: Sequence[str],
+        version_override: Optional[str] = None,
+    ):
+        EleutherTaskWithRenamedSplits.__init__(
+            self,
+            eleuther_task,
+            version_override=version_override,
+            ranked_classification=True)
+        WithAnswerOptionsMixin.__init__(self, answer_options)
+        self.add_metrics(classification_metrics(len(answer_options)))